For mercycorps/TolaActivity#2166, update and add tests.
