+++ js/pages/tola_management_pages/country/models.js
@@ -293,6 +293,10 @@ export class CountryStore {
             }
             this.countries.unshift(new_country_data)
             this.editing_target = 'new'
+            this.editing_disaggregations_data = [];
+            this.editing_objectives_data = [];
+            this.editing_errors = {};
+            this.editing_disaggregations_errors = {};
         }
     }
 
@@ -305,8 +309,6 @@ export class CountryStore {
                 this.saving = false;
                 this.editing_errors = {};
                 this.editing_target = response.data.id;
-                this.editing_disaggregations_data = [];
-                this.editing_objectives_data = [];
                 this.active_pane_is_dirty = false;
                 this.countries.shift();
                 this.countries.unshift(response.data);
